+++ src/plot_timeseries.py
@@ -52,9 +52,9 @@ class PlotTs():
             marker = self.marker
         self.initializeAxes()
 
-        if dates is None:
-            return
         self.prepareTsValues(dates=dates, ts_values=ts_values, ref_values=ref_values)
+        if self.dates is None:
+            return
 
         self.ax.plot(self.dates, self.plot_values, marker)
         if self.replicate_flag:
@@ -108,6 +108,7 @@ class PlotTs():
         self.setYlims(ax=ax)
         self.setYticks(ax=ax)
         self.setGrid(status=True, ax=ax)
+        self.ui.figure.tight_layout()
 
     def setGrid(self, status, ax=None):
         if not ax: